# src/main/java/src/main/java/com/viafluvial/srvusuario/adapters/in/web/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.viafluvial.srvusuario.adapters.in.web.api;

import org.springframework.format.annotation.DateTimeFormat;
import com.viafluvial.srvusuario.adapters.in.web.dto.ExistsResponseApi;
import java.time.LocalDateTime;
import com.viafluvial.srvusuario.adapters.in.web.dto.PagedUserResponseApi;
import java.util.UUID;
import com.viafluvial.srvusuario.adapters.in.web.dto.UserCreateDTOApi;
import com.viafluvial.srvusuario.adapters.in.web.dto.UserDTOApi;
import com.viafluvial.srvusuario.adapters.in.web.dto.UserStatusApi;
import com.viafluvial.srvusuario.adapters.in.web.dto.UserTypeApi;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-07T21:43:37.860179204Z[Etc/UTC]", comments = "Generator version: 7.6.0")
@Validated
@Tag(name = "Users", description = "the Users API")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /users : Create user
     *
     * @param userCreateDTOApi  (required)
     * @return User created (status code 201)
     *         or Invalid request (status code 400)
     *         or Email already exists (status code 409)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "User created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTOApi.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request"),
            @ApiResponse(responseCode = "409", description = "Email already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<UserDTOApi> createUser(
        @Parameter(name = "UserCreateDTOApi", description = "", required = true) @Valid @RequestBody UserCreateDTOApi userCreateDTOApi
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /users/{id} : Delete user
     *
     * @param id  (required)
     * @return User deleted (status code 204)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Delete user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "204", description = "User deleted"),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/users/{id}"
    )
    
    default ResponseEntity<Void> deleteUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/me : Get current user
     *
     * @return Current user (status code 200)
     *         or Unable to resolve user (status code 400)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "getCurrentUser",
        summary = "Get current user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Current user", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTOApi.class))
            }),
            @ApiResponse(responseCode = "400", description = "Unable to resolve user"),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/me",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserDTOApi> getCurrentUser(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/email/{email} : Get user by email
     *
     * @param email  (required)
     * @return User found (status code 200)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "getUserByEmail",
        summary = "Get user by email",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTOApi.class))
            }),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/email/{email}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserDTOApi> getUserByEmail(
        @Parameter(name = "email", description = "", required = true, in = ParameterIn.PATH) @PathVariable("email") String email
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/{id} : Get user by id
     *
     * @param id  (required)
     * @return User found (status code 200)
     *         or User not found (status code 404)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Get user by id",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTOApi.class))
            }),
            @ApiResponse(responseCode = "404", description = "User not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserDTOApi> getUserById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/type/{userType} : List users by type
     *
     * @param userType  (required)
     * @return Users by type (status code 200)
     */
    @Operation(
        operationId = "getUsersByType",
        summary = "List users by type",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Users by type", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserDTOApi.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/type/{userType}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<UserDTOApi>> getUsersByType(
        @Parameter(name = "userType", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userType") UserTypeApi userType
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users : Search users
     *
     * @param email  (optional)
     * @param type  (optional)
     * @param status  (optional)
     * @param emailVerified  (optional)
     * @param createdFrom  (optional)
     * @param createdTo  (optional)
     * @param page  (optional, default to 1)
     * @param size  (optional, default to 20)
     * @return Paged users (status code 200)
     */
    @Operation(
        operationId = "searchUsers",
        summary = "Search users",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Paged users", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PagedUserResponseApi.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PagedUserResponseApi> searchUsers(
        @Parameter(name = "email", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) String email,
        @Parameter(name = "type", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "type", required = false) UserTypeApi type,
        @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) UserStatusApi status,
        @Parameter(name = "emailVerified", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "emailVerified", required = false) Boolean emailVerified,
        @Parameter(name = "createdFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "createdFrom", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime createdFrom,
        @Parameter(name = "createdTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "createdTo", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime createdTo,
        @Min(1) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "1") Integer page,
        @Min(1) @Max(200) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"totalItems\" : 1, \"size\" : 6, \"totalPages\" : 5, \"page\" : 0, \"items\" : [ { \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /users/{id} : Update user
     *
     * @param id  (required)
     * @param userDTOApi  (required)
     * @return User updated (status code 200)
     *         or User not found (status code 404)
     *         or Invalid request (status code 400)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Update user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTOApi.class))
            }),
            @ApiResponse(responseCode = "404", description = "User not found"),
            @ApiResponse(responseCode = "400", description = "Invalid request")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/users/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<UserDTOApi> updateUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "UserDTOApi", description = "", required = true) @Valid @RequestBody UserDTOApi userDTOApi
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"emailVerified\" : true, \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastLogin\" : \"2000-01-23T04:56:07.000+00:00\", \"password\" : \"password\", \"phone\" : \"phone\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userType\" : \"PASSENGER\", \"email\" : \"email\", \"status\" : \"PENDING\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/exists : Check user existence by email
     *
     * @param email  (required)
     * @return Exists response (status code 200)
     */
    @Operation(
        operationId = "userExists",
        summary = "Check user existence by email",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Exists response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ExistsResponseApi.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/exists",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ExistsResponseApi> userExists(
        @NotNull @jakarta.validation.constraints.Email @Parameter(name = "email", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = true) String email
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"exists\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}