# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ViaFluvial User Service API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
  /users:
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDTO'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid request
        '409':
          description: Email already exists
    get:
      tags: [Users]
      summary: Search users
      operationId: searchUsers
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserType'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: emailVerified
          in: query
          required: false
          schema:
            type: boolean
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: Paged users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserResponse'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
        '400':
          description: Invalid request
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
  /users/email/{email}:
    get:
      tags: [Users]
      summary: Get user by email
      operationId: getUserByEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
  /users/type/{userType}:
    get:
      tags: [Users]
      summary: List users by type
      operationId: getUsersByType
      parameters:
        - name: userType
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        '200':
          description: Users by type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
  /users/exists:
    get:
      tags: [Users]
      summary: Check user existence by email
      operationId: userExists
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Exists response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExistsResponse'
  /users/me:
    get:
      tags: [Users]
      summary: Get current user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Unable to resolve user
        '404':
          description: User not found
...
components:
  schemas:
    UserType:
      type: string
      enum: [PASSENGER, BOATMAN, AGENCY, ADMIN]
    UserStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, INACTIVE]
...
    UserCreateDTO:
      type: object
      required: [userType, email, password, fullName, phone]
      properties:
        userType:
          $ref: '#/components/schemas/UserType'
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        phone:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userType:
          $ref: '#/components/schemas/UserType'
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        phone:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
...
    ExistsResponse:
      type: object
      properties:
        exists:
          type: boolean
    PagedUserResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDTO'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
...